+++ resources/displayFilterWindow.py
@@ -54,7 +54,6 @@ class displayFilterWindow(QWidget):
                 background-color: #f8f8f8;
                 border: 1px solid lightgray;
                 font-family: Courier New;
-                font-size: 12;
             }
         """)