+++ resources/displaySampleWindow.py
@@ -79,7 +79,6 @@ class displaySampleWindow(QWidget):
                 background-color: #f8f8f8;
                 border: 1px solid lightgray;
                 font-family: Courier New;
-                font-size: 12;
             }
         """)